feat: enhance affiliate management with DTOs, validation, and exception handling

+++ server/src/modules/affiliates/application/use-cases/delete-affiliate.use-case.ts
@@ -1,12 +1,20 @@
 import { PrismaAffiliateRepository } from "@/common/infrastructure/repositories";
+import { AffiliateNotFoundException } from "../../domain/exceptions/affiliate-not-found.exception";
+import { AffiliateEntity } from "@/common/domain/entites";
 
 export class DeleteAffiliateUseCase {
   constructor(
-    private repo =
-      new PrismaAffiliateRepository()
-  ) {}
+    private repo = new PrismaAffiliateRepository()
+  ) { }
 
-  execute(id: string) {
-    return this.repo.delete(id);
+  async execute(id: string): Promise<AffiliateEntity> {
+    const affiliate = await this.repo.findById(id);
+
+    if (!affiliate) {
+      throw new AffiliateNotFoundException(id);
+    }
+
+    const deleted = await this.repo.delete(id);
+    return deleted;
   }
 }
